# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: 后台服务接口
  version: 1.0.0
  title: BFFast快速开发平台
basePath: /fast
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/version/list/page:
    get:
      tags:
        - 系统管理-版本管理
      summary: 查看列表-分页
      produces:
        - application/json
      parameters:
        - name: current
          description: 页码
          in: query
          type: integer
          required: true
        - name: size
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: versionName
          description: 版本名称
          in: query
          type: string
          required: false
        - name: clientType
          description: 客户端类型 0-安卓 1-苹果
          in: query
          type: string
          required: false
      responses:
        200:
          description: 返回列表
          schema:
            $ref: '#/definitions/versionEntityListPage'

  #系统管理-版本管理-列表-不分页
...
  /sys/version/current/{clientType}:
    get:
      tags:
        - 系统管理-版本管理
      summary: 查询最新版本
      produces:
        - application/json
      parameters:
        - name: clientType
          description: 客户端类型 0-安卓 1-苹果
          in: path
          type: string
          required: true
      responses:
        200:
          description: 返回列表
          schema:
            $ref: '#/definitions/versionEntity'
  #系统管理-版本管理-保存
  /sys/version/save:
    post:
      tags:
        - 系统管理-版本管理
      summary: 保存
      produces:
        - application/json
      parameters:
        - name: body
          description: version 对象
          in: body
          schema:
            $ref: '#/definitions/versionEntity'
          required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/ResultDTO'

  #系统管理-版本管理-修改
  /sys/version/update:
    post:
      tags:
        - 系统管理-版本管理
      summary: 修改
      produces:
        - application/json
      parameters:
        - name: body
          description: version 对象
          in: body
          schema:
            $ref: '#/definitions/versionEntity'
          required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/ResultDTO'

  #系统管理-版本管理-单条查询
  /sys/version/info/{id}:
    get:
      tags:
        - 系统管理-版本管理
      summary: 获取单条记录
      produces:
        - application/json
      parameters:
        - name: id
          description: 记录ID
          in: path
          type: integer
          required: true
      responses:
        200:
          description: 返回单条记录
          schema:
            type: object
            properties:
              code:
                description: 状态码 0-成功  非0-失败
                type: integer
                format: int32
              msg:
                description: 返回结果消息
                type: string
              data:
                $ref: '#/definitions/versionEntity'

  #系统管理-版本管理-删除
  /sys/version/delete:
    post:
      tags:
        - 系统管理-版本管理
      summary: 删除
      produces:
        - application/json
      parameters:
        - name: body
          description: ID列表
          in: body
          schema:
            type: array
            items:
              type: integer
              format: int64
              default: 0
          required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/ResultDTO'
  #系统管理-开放平台-列表-分页
...
definitions:
  ResultDTO:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
        default: 0
      msg:
        description: 失败原因
        type: string
        default: success
      data:
        description: 数据对象
        type: object
...
  versionEntityListPage:
    type: object
    properties:
      code:
        description: 状态码 0-成功  非0-失败
        type: integer
        format: int32
      msg:
        description: 返回结果消息
        type: string
      data:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/versionEntity'
  #系统管理-版本管理-EntityList
...
  versionEntity:
    type: object
    properties:
      id:
        description: id
        type: integer
        format: int64
      clientType:
        description: 客户端类型 0-安卓 1-苹果
        type: string
      versionName:
        description: 版本名称
        type: string
      version:
        description: 版本号
        type: integer
        format: int32
      content:
        description: 描述
        type: string
      upgrade:
        description: 0-不强制,1-强制升级
        type: string
      isCurrent:
        description: 是否最新版本 0-否 1-是
        type: string
      installUrl:
        description: 安装路径
        type: string
      createTime:
        description: 创建时间
        type: integer
        format: int64
      createUser:
        description: 创建人
        type: integer
        format: int64
      createUserName:
        description: 创建人名称
        type: string
...